VKBD cleanup
