+++ libretro/libretro-core.h
@@ -45,6 +45,7 @@ extern struct zfile *retro_deserialize_file;
 extern dc_storage *retro_dc;
 extern retro_log_printf_t log_cb;
 extern long retro_ticks(void);
+extern int RGBc(int r, int g, int b);
 extern int umain (int argc, TCHAR **argv);
 
 /* Statusbar */
